+++ modules/interpreter/src/include/Scope.hpp
@@ -272,6 +272,9 @@ public:
 
     bool
     isVariable(const std::string& varname);
+
+    void
+    clearCache();
 };
 //=============================================================================
 } // namespace Nelson